new directory structure: merge config to media, move everything up in scope.
new struct: media/config
it is now a subfolder of media (this is a intermediate solution, but since we need to remake the "semi-hardcoded" solutions in config anyways this is the one which makes the most sense for now. We however need to move this stuff from here to their respective media folders.).

Furthermore since we only had a single folder in here, we can finally get rid of "media/" but make it an extension to the main folder (done with the addpackagedir-cmd).